products eklenme aşaması

+++ src/ElectronicProduct.java
@@ -2,7 +2,7 @@ public class ElectronicProduct extends BaseProduct {
     ElectronicProduct(){
         setCatagoryId(1);
     }
-    ElectronicProduct(int price,String productCompany,String productName){
+    ElectronicProduct(double price,String productCompany,String productName){
         setCatagoryId(1);
         this.price=price;
         this.productCompany=productCompany;
