+++ src/HouseholdAppliancesProduct.java
@@ -3,7 +3,7 @@ public class HouseholdAppliancesProduct extends BaseProduct{
         setCatagoryId(3);
     }
 
-    HouseholdAppliancesProduct(int price, String productCompany, String productName) {
+    HouseholdAppliancesProduct(double price, String productCompany, String productName) {
         setCatagoryId(3);
         this.price = price;
         this.productCompany = productCompany;
